setup-server/upload/requirements.txt
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
opencv-contrib-python==4.5.3.56
Pillow==8.0.1
pymongo==4.0.1
python-bsonjs==0.2.0
requests==2.24.0
tensorflow==2.6.0
urllib3==1.25.11
SQLAlchemy==1.3.18
Werkzeug==1.0.1
Jinja2==2.11.2
MarkupSafe==1.1.1
itsdangerous==1.1.0
